requirements.txt
django
pyTelegramBotAPI
commentjson
imageio
numpy
opencv-python-headless
pybind11
pyquaternion
scipy
tqdm
pycolmap<=0.4.0,>=0.3.0
aiogram==2.23.1
python-dotenv
requests